Fixes bug in migration.
Adds rake file_tracker:reindex to re-index all tracked files.

+++ db/migrate/20171205221900_remove_duracloud_columns.rb
@@ -1,6 +1,6 @@
 class RemoveDuracloudColumns < ActiveRecord::Migration[5.1]
   def up
-    drop_table :duracloud_manifest_entries
+    drop_table :duracloud_manifest_entries if table_exists?(:duracloud_manifest_entries)
     remove_columns :tracked_directories, :duracloud_space, :duracloud_checked_at
     remove_columns :tracked_files, :duracloud_status, :duracloud_checked_at
   end
